pyproject.toml
[tool.poetry]
name = "at-better-web"
version = "0.1.0"
description = ""
authors = ["lele <lele@datamini.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
# 原有依赖
python = "^3.11"
asktable = "^4.2.0"
pydantic-settings = "^2.7.0"
pybase62 = "^1.0.0"
sqlalchemy = "^2.0.36"
requests = "^2.32.3"
pyyaml = "^6.0.2"
pymysql = "^1.1.1"
tabulate = "^0.9.0"
lark-oapi = "^1.4.6"

# 新增依赖
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
jinja2 = "^3.1.2"

[tool.poetry.scripts]
# 转换为脚本
at-better-web = "app.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
